backend/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.34.0
python-socketio==5.12.1
python-multipart==0.0.20
jsonschema==3.2.0
lxml==6.0.2
numpy==2.3.4
matplotlib==3.10.5
Requests==2.32.4
Werkzeug==3.1.3
pymoose==4.1.3
prometheus-fastapi-instrumentator
aiofiles==24.1.0
httpx